src/test/java/com/aegis/query/QueryExecutorTest.java
package com.aegis.query;

import com.aegis.domain.QueryResult;
import com.aegis.domain.StorageTier;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.opensearch.client.RestHighLevelClient;
import org.springframework.jdbc.core.JdbcTemplate;
import reactor.core.publisher.Flux;
import reactor.test.StepVerifier;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static org.assertj.core.api.Assertions.assertThat;

/**
 * Unit tests for QueryExecutor
 * Tests the query execution interface and client dependencies
 */
@ExtendWith(MockitoExtension.class)
class QueryExecutorTest {
    
    @Mock
    private RestHighLevelClient openSearchClient;
    
    @Mock
    private JdbcTemplate clickHouseTemplate;
    
    private QueryExecutor queryExecutor;
    
    @BeforeEach
    void setUp() {
        queryExecutor = new QueryExecutor(openSearchClient, clickHouseTemplate);
    }
    
    @Test
    void testConstructor_WithValidDependencies_ShouldCreateInstance() {
        // Given: Valid client dependencies
        // When: QueryExecutor is created
        // Then: Instance should be created successfully
        assertThat(queryExecutor).isNotNull();
    }
    
    @Test
    void testExecute_WithNullPlan_ShouldReturnEmptyResult() {
        // Given: A null query plan
        QueryPlan plan = null;
        
        // When: Execute is called
        Flux<QueryResult> result = queryExecutor.execute(plan);
        
        // Then: Should return empty result
        StepVerifier.create(result)
            .assertNext(queryResult -> {
                assertThat(queryResult).isNotNull();
                assertThat(queryResult.getRows()).isEmpty();
            })
            .verifyComplete();
    }
    
    @Test
    void testExecute_WithEmptyPlan_ShouldReturnEmptyResult() {
        // Given: An empty query plan
        QueryPlan plan = new QueryPlan();
        
        // When: Execute is called
        Flux<QueryResult> result = queryExecutor.execute(plan);
        
        // Then: Should return empty result
        StepVerifier.create(result)
            .assertNext(queryResult -> {
                assertThat(queryResult).isNotNull();
                assertThat(queryResult.getRows()).isEmpty();
            })
            .verifyComplete();
    }
    
    @Test
    void testExecute_WithSingleHotTierQuery_ShouldExecuteSuccessfully() {
        // Given: A query plan with a single hot tier query
        QueryPlan plan = new QueryPlan();
        OpenSearchQuery hotQuery = new OpenSearchQuery(null); // Placeholder
        plan.addSubQuery(hotQuery);
        
        // When: Execute is called
        Flux<QueryResult> result = queryExecutor.execute(plan);
        
        // Then: Should return a result (placeholder implementation returns empty)
        StepVerifier.create(result)
            .assertNext(queryResult -> {
                assertThat(queryResult).isNotNull();
            })
            .verifyComplete();
    }
    
    @Test
    void testExecute_WithSingleWarmTierQuery_ShouldExecuteSuccessfully() {
        // Given: A query plan with a single warm tier query
        QueryPlan plan = new QueryPlan();
        ClickHouseQuery warmQuery = new ClickHouseQuery("SELECT * FROM aegis_events_warm");
        plan.addSubQuery(warmQuery);
        
        // When: Execute is called
        Flux<QueryResult> result = queryExecutor.execute(plan);
        
        // Then: Should return a result (placeholder implementation returns empty)
        StepVerifier.create(result)
            .assertNext(queryResult -> {
                assertThat(queryResult).isNotNull();
            })
            .verifyComplete();
    }
    
    @Test
    void testExecute_WithMultipleTierQueries_ShouldMergeResults() {
        // Given: A query plan with queries for multiple tiers
        QueryPlan plan = new QueryPlan();
        OpenSearchQuery hotQuery = new OpenSearchQuery(null);
        ClickHouseQuery warmQuery = new ClickHouseQuery("SELECT * FROM aegis_events_warm");
        plan.addSubQuery(hotQuery);
        plan.addSubQuery(warmQuery);
        
        // When: Execute is called
        Flux<QueryResult> result = queryExecutor.execute(plan);
        
        // Then: Should merge results from both tiers
        StepVerifier.create(result)
            .assertNext(queryResult -> {
                assertThat(queryResult).isNotNull();
                // Results will be merged (currently both return empty)
            })
            .verifyComplete();
    }
    
    @Test
    void testMergeResults_WithValidResults_ShouldCombineRows() {
        // This test verifies the merge logic indirectly through execute
        // Direct testing of mergeResults will be added when the method is enhanced
        
        // Given: Multiple query results
        QueryResult result1 = new QueryResult();
        Map<String, Object> row1 = new HashMap<>();
        row1.put("id", "1");
        row1.put("message", "Event 1");
        result1.addRow(row1);
        result1.setTotalCount(1);
        
        QueryResult result2 = new QueryResult();
        Map<String, Object> row2 = new HashMap<>();
        row2.put("id", "2");
        row2.put("message", "Event 2");
        result2.addRow(row2);
        result2.setTotalCount(1);
        
        // When: Results are merged (simulated)
        QueryResult merged = new QueryResult();
        merged.addAll(result1.getRows());
        merged.addAll(result2.getRows());
        merged.setTotalCount(result1.getTotalCount() + result2.getTotalCount());
        
        // Then: Should contain rows from both results
        assertThat(merged.getRows()).hasSize(2);
        assertThat(merged.getTotalCount()).isEqualTo(2);
    }
    
    @Test
    void testQueryPlanIntegration_WithComplexPlan_ShouldHandleCorrectly() {
        // Given: A complex query plan with multiple sub-queries
        QueryPlan plan = new QueryPlan();
        
        // Add queries for different tiers
        plan.addSubQuery(new OpenSearchQuery(null));
        plan.addSubQuery(new ClickHouseQuery("SELECT COUNT(*) FROM aegis_events_warm"));
        
        // When: Execute is called
        Flux<QueryResult> result = queryExecutor.execute(plan);
        
        // Then: Should handle all sub-queries
        StepVerifier.create(result)
            .assertNext(queryResult -> {
                assertThat(queryResult).isNotNull();
                assertThat(plan.size()).isEqualTo(2);
            })
            .verifyComplete();
    }
}
